Create units.utils.kubevirt with common stuff
